requirements.txt
mysql-connector-python
transformers
pyspark 
pandas
torch
numpy
Flask
sklearn